trekk logikk som ikke hører hjemme i SmåbarnstilleggUtils ut i mer passende filer
