+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/beregning/SmåbarnstilleggUtils.kt
@@ -1,14 +1,8 @@
 package no.nav.familie.ba.sak.kjerne.beregning
 
 import no.nav.familie.ba.sak.common.Feil
-import no.nav.familie.ba.sak.common.MånedPeriode
-import no.nav.familie.ba.sak.common.isSameOrAfter
-import no.nav.familie.ba.sak.common.toYearMonth
-import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelse
 import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.InternPeriodeOvergangsstønad
-import no.nav.familie.ba.sak.kjerne.beregning.domene.YtelseType
-import no.nav.familie.ba.sak.kjerne.forrigebehandling.EndringIUtbetalingUtil
 import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.ba.sak.kjerne.tidslinjefamiliefelles.komposisjon.erTilogMed3ÅrTidslinje
 import no.nav.familie.ba.sak.kjerne.tidslinjefamiliefelles.komposisjon.kombinerUtenNull
@@ -17,98 +11,8 @@ import no.nav.familie.tidslinje.Tidslinje
 import no.nav.familie.tidslinje.beskjærEtter
 import no.nav.familie.tidslinje.utvidelser.filtrerIkkeNull
 import no.nav.familie.tidslinje.utvidelser.kombinerMed
-import no.nav.familie.tidslinje.utvidelser.tilPerioder
-import no.nav.familie.tidslinje.utvidelser.tilPerioderIkkeNull
-import org.springframework.http.HttpStatus
 import java.math.BigDecimal
 import java.time.LocalDate
-import java.time.YearMonth
-
-class VedtaksperiodefinnerSmåbarnstilleggFeil(
-    melding: String,
-    override val frontendFeilmelding: String? = null,
-    override val httpStatus: HttpStatus = HttpStatus.OK,
-    override val throwable: Throwable? = null,
-) : Feil(
-        melding,
-        frontendFeilmelding,
-        httpStatus,
-        throwable,
-    )
-
-fun vedtakOmOvergangsstønadPåvirkerFagsak(
-    småbarnstilleggGenerator: SmåbarnstilleggGenerator,
-    nyePerioderMedFullOvergangsstønad: List<InternPeriodeOvergangsstønad>,
-    forrigeAndelerTilkjentYtelse: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
-    barnasAktørerOgFødselsdatoer: List<Pair<Aktør, LocalDate>>,
-): Boolean {
-    val (forrigeSmåbarnstilleggAndeler, forrigeAndelerIkkeSmåbarnstillegg) = forrigeAndelerTilkjentYtelse.partition { it.erSmåbarnstillegg() }
-
-    val (forrigeUtvidetAndeler, forrigeBarnasAndeler) = forrigeAndelerIkkeSmåbarnstillegg.partition { it.erUtvidet() }
-
-    val nyeSmåbarnstilleggAndeler =
-        småbarnstilleggGenerator.lagSmåbarnstilleggAndeler(
-            perioderMedFullOvergangsstønad = nyePerioderMedFullOvergangsstønad,
-            barnasAndeler = forrigeBarnasAndeler,
-            utvidetAndeler = forrigeUtvidetAndeler,
-            barnasAktørerOgFødselsdatoer = barnasAktørerOgFødselsdatoer,
-        )
-
-    return nyeSmåbarnstilleggAndeler.førerTilEndringIUtbetalingFraForrigeBehandling(
-        forrigeAndeler = forrigeSmåbarnstilleggAndeler,
-    )
-}
-
-private fun List<AndelTilkjentYtelseMedEndreteUtbetalinger>.førerTilEndringIUtbetalingFraForrigeBehandling(
-    forrigeAndeler: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
-): Boolean {
-    val endringstidslinje =
-        EndringIUtbetalingUtil.lagEndringIUtbetalingTidslinje(
-            nåværendeAndeler = this.map { it.andel },
-            forrigeAndeler = forrigeAndeler.map { it.andel },
-        )
-
-    return endringstidslinje.tilPerioder().any { it.verdi == true }
-}
-
-fun hentInnvilgedeOgReduserteAndelerSmåbarnstillegg(
-    forrigeSmåbarnstilleggAndeler: List<AndelTilkjentYtelse>,
-    nyeSmåbarnstilleggAndeler: List<AndelTilkjentYtelse>,
-): Pair<List<MånedPeriode>, List<MånedPeriode>> {
-    val forrigeAndelerTidslinje = forrigeSmåbarnstilleggAndeler.tilTidslinjeForSøkersYtelse(YtelseType.SMÅBARNSTILLEGG)
-    val andelerTidslinje = nyeSmåbarnstilleggAndeler.tilTidslinjeForSøkersYtelse(YtelseType.SMÅBARNSTILLEGG)
-
-    val nyeSmåbarnstilleggPerioder =
-        forrigeAndelerTidslinje.kombinerMed(andelerTidslinje) { gammel, ny -> ny.takeIf { gammel == null } }
-
-    val fjernedeSmåbarnstilleggPerioder =
-        forrigeAndelerTidslinje.kombinerMed(andelerTidslinje) { gammel, ny -> gammel.takeIf { ny == null } }
-
-    return Pair(nyeSmåbarnstilleggPerioder.tilMånedPerioder(), fjernedeSmåbarnstilleggPerioder.tilMånedPerioder())
-}
-
-private fun Tidslinje<AndelTilkjentYtelse>.tilMånedPerioder() =
-    this.tilPerioderIkkeNull().map {
-        MånedPeriode(
-            fom = it.fom?.toYearMonth() ?: throw Feil("Fra og med-dato kan ikke være null"),
-            tom = it.tom?.toYearMonth() ?: throw Feil("Til og med-dato kan ikke være null"),
-        )
-    }
-
-fun kanAutomatiskIverksetteSmåbarnstillegg(
-    innvilgedeMånedPerioder: List<MånedPeriode>,
-    reduserteMånedPerioder: List<MånedPeriode>,
-): Boolean =
-    innvilgedeMånedPerioder.all {
-        it.fom.isSameOrAfter(
-            YearMonth.now(),
-        )
-    } &&
-        reduserteMånedPerioder.all {
-            it.fom.isSameOrAfter(
-                YearMonth.now(),
-            )
-        }
 
 fun kombinerBarnasTidslinjerTilUnder3ÅrResultat(
     alleAndelerForBarnUnder3År: Iterable<AndelTilkjentYtelseMedEndreteUtbetalinger>,
