+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/grunnlag/overgangsstønad/OvergangsstønadUtils.kt
@@ -1,10 +1,14 @@
 package no.nav.familie.ba.sak.kjerne.grunnlag.overgangsstønad
 
 import no.nav.familie.ba.sak.common.toYearMonth
+import no.nav.familie.ba.sak.kjerne.beregning.SmåbarnstilleggGenerator
+import no.nav.familie.ba.sak.kjerne.beregning.domene.AndelTilkjentYtelseMedEndreteUtbetalinger
 import no.nav.familie.ba.sak.kjerne.beregning.domene.InternPeriodeOvergangsstønad
 import no.nav.familie.ba.sak.kjerne.beregning.domene.slåSammenTidligerePerioder
 import no.nav.familie.ba.sak.kjerne.beregning.domene.splitFramtidigePerioderFraForrigeBehandling
 import no.nav.familie.ba.sak.kjerne.beregning.domene.tilTidslinje
+import no.nav.familie.ba.sak.kjerne.forrigebehandling.EndringIUtbetalingUtil
+import no.nav.familie.ba.sak.kjerne.personident.Aktør
 import no.nav.familie.kontrakter.felles.ef.EksternPeriode
 import no.nav.familie.tidslinje.Periode
 import no.nav.familie.tidslinje.Tidslinje
@@ -12,6 +16,42 @@ import no.nav.familie.tidslinje.tilTidslinje
 import no.nav.familie.tidslinje.utvidelser.kombinerMed
 import no.nav.familie.tidslinje.utvidelser.tilPerioder
 import java.time.LocalDate
+
+fun vedtakOmOvergangsstønadPåvirkerFagsak(
+    småbarnstilleggGenerator: SmåbarnstilleggGenerator,
+    nyePerioderMedFullOvergangsstønad: List<InternPeriodeOvergangsstønad>,
+    forrigeAndelerTilkjentYtelse: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
+    barnasAktørerOgFødselsdatoer: List<Pair<Aktør, LocalDate>>,
+): Boolean {
+    val (forrigeSmåbarnstilleggAndeler, forrigeAndelerIkkeSmåbarnstillegg) = forrigeAndelerTilkjentYtelse.partition { it.erSmåbarnstillegg() }
+
+    val (forrigeUtvidetAndeler, forrigeBarnasAndeler) = forrigeAndelerIkkeSmåbarnstillegg.partition { it.erUtvidet() }
+
+    val nyeSmåbarnstilleggAndeler =
+        småbarnstilleggGenerator.lagSmåbarnstilleggAndeler(
+            perioderMedFullOvergangsstønad = nyePerioderMedFullOvergangsstønad,
+            barnasAndeler = forrigeBarnasAndeler,
+            utvidetAndeler = forrigeUtvidetAndeler,
+            barnasAktørerOgFødselsdatoer = barnasAktørerOgFødselsdatoer,
+        )
+
+    return nyeSmåbarnstilleggAndeler.førerTilEndringIUtbetalingFraForrigeBehandling(
+        forrigeAndeler = forrigeSmåbarnstilleggAndeler,
+    )
+}
+
+
+private fun List<AndelTilkjentYtelseMedEndreteUtbetalinger>.førerTilEndringIUtbetalingFraForrigeBehandling(
+    forrigeAndeler: List<AndelTilkjentYtelseMedEndreteUtbetalinger>,
+): Boolean {
+    val endringstidslinje =
+        EndringIUtbetalingUtil.lagEndringIUtbetalingTidslinje(
+            nåværendeAndeler = this.map { it.andel },
+            forrigeAndeler = forrigeAndeler.map { it.andel },
+        )
+
+    return endringstidslinje.tilPerioder().any { it.verdi == true }
+}
 
 fun List<InternPeriodeOvergangsstønad>.splittOgSlåSammen(
     overgangsstønadPerioderFraForrigeBehandling: List<InternPeriodeOvergangsstønad>,
